+++ lib/gds_api/json_client.rb
@@ -52,8 +52,8 @@ module GdsApi
       do_json_request(:get, url, nil, additional_headers, &create_response)
     end
 
-    def post_json(url, params = {}, additional_headers = {})
-      do_json_request(:post, url, params, additional_headers)
+    def post_json(url, params = {}, additional_headers = {}, &create_response)
+      do_json_request(:post, url, params, additional_headers, &create_response)
     end
 
     def put_json(url, params, additional_headers = {})